Introduce the system property "appengine.use.EE10" to force the local devappserver execution to use Jetty12 EE10 profile. WIP. still does not compile. Remove unused code (ResourceFileServlet and TransactionCleanupListener I think).
