+++ api_dev/src/main/java/com/google/appengine/tools/development/ContainerService.java
@@ -20,11 +20,7 @@ import com.google.appengine.tools.development.ApplicationConfigurationManager.Mo
 import com.google.apphosting.api.ApiProxy;
 import com.google.apphosting.utils.config.AppEngineWebXml;
 import java.io.File;
-import java.io.IOException;
 import java.util.Map;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
  * Provides the backing servlet container support for the {@link DevAppServer},
@@ -131,17 +127,4 @@ public interface ContainerService {
    */
   Map<String, String> getServiceProperties();
 
-  /**
-   * Forwards an HttpRequest request to this container.
-   */
-  void forwardToServer(HttpServletRequest hrequest, HttpServletResponse hresponse)
-      throws IOException, ServletException;
-
-  /** Forwards an EE10 HttpRequest request to this container. */
-  default void forwardToServerEE10(
-      jakarta.servlet.http.HttpServletRequest hrequest,
-      jakarta.servlet.http.HttpServletResponse hresponse)
-      throws IOException, ServletException {
-      
-  }
 }
