+++ api_dev/src/main/java/com/google/appengine/tools/development/DelegatingModulesFilterHelperEE8.java
@@ -1,0 +1,41 @@
+
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.appengine.tools.development;
+
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+/**
+ *
+ */
+public class DelegatingModulesFilterHelperEE8 extends DelegatingModulesFilterHelper {
+    
+  public DelegatingModulesFilterHelperEE8(BackendServers backendServers, Modules modules) {
+    super(backendServers, modules);
+  }
+  public void forwardToInstance(String moduleOrBackendName, int instance,
+      HttpServletRequest hrequest, HttpServletResponse response)
+      throws IOException, ServletException {
+      if (isBackend(moduleOrBackendName)) {
+        ((BackendServersEE8)backendServers).forwardToServer(moduleOrBackendName, instance, hrequest, response);
+     } else {
+       ((ModulesEE8)modules).forwardToInstance(moduleOrBackendName, instance, hrequest, response);
+     }
+  }    
+}
