+++ api_dev/src/main/java/com/google/appengine/tools/development/ee10/DevAppServerModulesFilter.java
@@ -17,22 +17,15 @@
 package com.google.appengine.tools.development.ee10;
 
 import com.google.appengine.api.backends.BackendService;
-import com.google.appengine.api.backends.dev.LocalServerController;
-import com.google.appengine.api.modules.ModulesException;
 import com.google.appengine.api.modules.ModulesService;
 import com.google.appengine.api.modules.ModulesServiceFactory;
-import com.google.appengine.tools.development.AbstractBackendServers;
 import com.google.appengine.tools.development.BackendServers;
-import com.google.appengine.tools.development.DevAppServerImpl;
+import com.google.appengine.tools.development.DevAppServerModulesCommon;
 import com.google.appengine.tools.development.LocalEnvironment;
-import com.google.appengine.tools.development.Modules;
-import com.google.appengine.tools.development.ModulesController;
+import com.google.appengine.tools.development.ModulesFilterHelper;
 import com.google.apphosting.api.ApiProxy;
 import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
-import java.util.Map;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.FilterConfig;
@@ -83,14 +76,7 @@ import jakarta.servlet.http.HttpServletResponse;
  *
  *
  */
-public class DevAppServerModulesFilter implements Filter {
-
-  static final String BACKEND_REDIRECT_ATTRIBUTE = "com.google.appengine.backend.BackendName";
-  static final String BACKEND_INSTANCE_REDIRECT_ATTRIBUTE =
-      "com.google.appengine.backend.BackendInstance";
-  @VisibleForTesting
-  static final String MODULE_INSTANCE_REDIRECT_ATTRIBUTE =
-      "com.google.appengine.module.ModuleInstance";
+public class DevAppServerModulesFilter extends DevAppServerModulesCommon implements Filter {
 
   // In prod instances return 500 (Internal Server Error) when busy
   static final int INSTANCE_BUSY_ERROR_CODE = HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
@@ -100,15 +86,9 @@ public class DevAppServerModulesFilter implements Filter {
 
   static final int MODULE_MISSING_ERROR_CODE = HttpServletResponse.SC_BAD_GATEWAY;
 
-  private final AbstractBackendServers backendServersManager;
-  private final ModulesService modulesService;
-
-  private final Logger logger = Logger.getLogger(DevAppServerModulesFilter.class.getName());
-
   @VisibleForTesting
-  DevAppServerModulesFilter(AbstractBackendServers backendServers, ModulesService modulesService) {
-    this.backendServersManager = backendServers;
-    this.modulesService = modulesService;
+  DevAppServerModulesFilter(BackendServers backendServers, ModulesService modulesService) {
+    super(backendServers, modulesService);
   }
 
   public DevAppServerModulesFilter() {
@@ -167,76 +147,37 @@ public class DevAppServerModulesFilter implements Filter {
     // a forwarded redirected request will have both sets of headers.
     if (hrequest.getRequestURI().equals("/_ah/start") &&
         expectsGeneratedStartRequests(backendServerName, instancePort)) {
-      return RequestType.STARTUP_REQUEST;
+      return DevAppServerModulesCommon.RequestType.STARTUP_REQUEST;
     } else if (hrequest.getAttribute(BACKEND_REDIRECT_ATTRIBUTE) instanceof String) {
       // this request was redirected here from a different instance
-      return RequestType.REDIRECTED_BACKEND_REQUEST;
+      return DevAppServerModulesCommon.RequestType.REDIRECTED_BACKEND_REQUEST;
     } else if (hrequest.getAttribute(MODULE_INSTANCE_REDIRECT_ATTRIBUTE) instanceof Integer) {
       // this request was redirected here from a different instance
-      return RequestType.REDIRECTED_MODULE_REQUEST;
+      return DevAppServerModulesCommon.RequestType.REDIRECTED_MODULE_REQUEST;
     } else if (backendServerName != null) {
       // request was to a backend server, check out if replica was specified
       int backendInstance = backendServersManager.getServerInstanceFromPort(instancePort);
       if (backendInstance == -1) {
         // no replica specified, redirect needed
-        return RequestType.REDIRECT_REQUESTED;
+        return DevAppServerModulesCommon.RequestType.REDIRECT_REQUESTED;
       } else {
-        return RequestType.DIRECT_BACKEND_REQUEST;
+        return DevAppServerModulesCommon.RequestType.DIRECT_BACKEND_REQUEST;
       }
     } else {
       // request to a non-backend instance, check if the user want us to redirect
       String serverRedirectHeader =
           getHeaderOrParameter(hrequest, BackendService.REQUEST_HEADER_BACKEND_REDIRECT);
       if (serverRedirectHeader == null && !isLoadBalancingRequest()) {
-        return RequestType.DIRECT_MODULE_REQUEST;
+        return DevAppServerModulesCommon.RequestType.DIRECT_MODULE_REQUEST;
       } else {
-        return RequestType.REDIRECT_REQUESTED;
+        return DevAppServerModulesCommon.RequestType.REDIRECT_REQUESTED;
       }
     }
   }
-
-  private boolean isLoadBalancingRequest() {
-    ModulesFilterHelper modulesFilterHelper = getModulesFilterHelper();
-    String module = modulesService.getCurrentModule();
-    int instance = getCurrentModuleInstance();
-    return modulesFilterHelper.isLoadBalancingInstance(module, instance);
-  }
-
-  private boolean expectsGeneratedStartRequests(String backendName,
-      int requestPort) {
-    String moduleOrBackendName = backendName;
-    if (moduleOrBackendName == null) {
-      moduleOrBackendName = modulesService.getCurrentModule();
-    }
-
-    int instance = backendName == null ? getCurrentModuleInstance() :
-      backendServersManager.getServerInstanceFromPort(requestPort);
-    ModulesFilterHelper modulesFilterHelper = getModulesFilterHelper();
-    return modulesFilterHelper.expectsGeneratedStartRequests(moduleOrBackendName, instance);
-  }
-
-  /**
-   * Returns the instance id for the module instance handling the current request or -1
-   * if a back end server or load balancing server is handling the request.
-   */
-  private int getCurrentModuleInstance() {
-    String instance = "-1";
-    try {
-      instance = modulesService.getCurrentInstanceId();
-    } catch (ModulesException me) {
-      logger.log(Level.FINEST, "Ignoring Exception getting module instance and continuing", me);
-    }
-    return Integer.parseInt(instance);
-  }
-
-  private ModulesFilterHelper getModulesFilterHelper() {
-    Map<String, Object> attributes = ApiProxy.getCurrentEnvironment().getAttributes();
-    return (ModulesFilterHelper) attributes.get(DevAppServerImpl.MODULES_FILTER_HELPER_PROPERTY);
-  }
-
+  
   private boolean tryToAcquireServingPermit(
       String moduleOrBackendName, int instance, HttpServletResponse hresponse) throws IOException {
-    ModulesFilterHelper modulesFilterHelper = getModulesFilterHelper();
+    ModulesFilterHelperEE10 modulesFilterHelper = (ModulesFilterHelperEE10) getModulesFilterHelper();
     // Instance specified, check if exists.
     if (!modulesFilterHelper.checkInstanceExists(moduleOrBackendName, instance)) {
       String msg =
@@ -293,7 +234,7 @@ public class DevAppServerModulesFilter implements Filter {
       moduleOrBackendName = modulesService.getCurrentModule();
       isLoadBalancingModuleInstance = true;
     }
-    ModulesFilterHelper modulesFilterHelper = getModulesFilterHelper();
+    ModulesFilterHelperEE10 modulesFilterHelper = (ModulesFilterHelperEE10)getModulesFilterHelper();
     int instance = getInstanceIdFromRequest(hrequest);
     logger.finest(String.format("redirect request to module: %d.%s", instance,
         moduleOrBackendName));
@@ -453,51 +394,6 @@ public class DevAppServerModulesFilter implements Filter {
   }
 
   /**
-   * Inject information about the current backend server setup so it is available
-   * to the BackendService API. This information is stored in the threadLocalAttributes
-   * in the current environment.
-   *
-   * @param backendName The server that is handling the request
-   * @param instance The server instance that is handling the request
-   */
-  private void injectApiInfo(String backendName, int instance) {
-    Map<String, String> portMapping = backendServersManager.getPortMapping();
-    if (portMapping == null) {
-      throw new IllegalStateException("backendServersManager.getPortMapping() is null");
-    }
-    injectBackendServiceCurrentApiInfo(backendName, instance, portMapping);
-
-    Map<String, Object> threadLocalAttributes = ApiProxy.getCurrentEnvironment().getAttributes();
-
-    // We inject backendServersManager which is not injected by
-    // injectBackendServiceCurrentApiInfo as it is not needed by BackendsService
-    // but is needed by the admin console for handling HTTP requests.
-    if (!portMapping.isEmpty()) {
-      threadLocalAttributes.put(
-          LocalServerController.BACKEND_CONTROLLER_ATTRIBUTE_KEY, backendServersManager);
-    }
-
-    threadLocalAttributes.put(
-        ModulesController.MODULES_CONTROLLER_ATTRIBUTE_KEY,
-        Modules.getInstance());
-  }
-
-  /**
-   * Sets up {@link ApiProxy} attributes needed {@link BackendService}.
-   */
-  public static void injectBackendServiceCurrentApiInfo(
-      String backendName, int backendInstance, Map<String, String> portMapping) {
-    Map<String, Object> threadLocalAttributes = ApiProxy.getCurrentEnvironment().getAttributes();
-    if (backendInstance != -1) {
-      threadLocalAttributes.put(BackendService.INSTANCE_ID_ENV_ATTRIBUTE, backendInstance + "");
-    }
-    if (backendName != null) {
-      threadLocalAttributes.put(BackendService.BACKEND_ID_ENV_ATTRIBUTE, backendName);
-    }
-    threadLocalAttributes.put(BackendService.DEVAPPSERVER_PORTMAPPING_KEY, portMapping);
-  }
-
-  /**
    * Checks the request headers and request parameters for the specified key
    */
   @VisibleForTesting
@@ -528,11 +424,5 @@ public class DevAppServerModulesFilter implements Filter {
     } catch (NumberFormatException e) {
       return -1;
     }
-  }
-
-  @VisibleForTesting
-  static enum RequestType {
-    DIRECT_MODULE_REQUEST, REDIRECT_REQUESTED, DIRECT_BACKEND_REQUEST, REDIRECTED_BACKEND_REQUEST,
-    REDIRECTED_MODULE_REQUEST, STARTUP_REQUEST;
   }
 }
