+++ api_dev/src/test/java/com/google/appengine/tools/development/BackendServersTest.java
@@ -16,7 +16,7 @@
 
 package com.google.appengine.tools.development;
 
-import com.google.appengine.tools.development.AbstractBackendServers.ServerInstanceEntry;
+import com.google.appengine.tools.development.BackendServers.ServerInstanceEntry;
 import java.util.HashSet;
 import junit.framework.TestCase;
