+++ runtime/local_jetty9/src/main/java/com/google/appengine/tools/development/jetty9/JettyContainerService.java
@@ -24,6 +24,7 @@ import com.google.appengine.tools.development.AbstractContainerService;
 import com.google.appengine.tools.development.ApiProxyLocal;
 import com.google.appengine.tools.development.AppContext;
 import com.google.appengine.tools.development.ContainerService;
+import com.google.appengine.tools.development.ContainerServiceEE8;
 import com.google.appengine.tools.development.DevAppServer;
 import com.google.appengine.tools.development.DevAppServerModulesFilter;
 import com.google.appengine.tools.development.IsolatedAppClassLoader;
@@ -75,7 +76,7 @@ import org.eclipse.jetty.webapp.WebAppContext;
  * Implements a Jetty backed {@link ContainerService}.
  *
  */
-public class JettyContainerService extends AbstractContainerService {
+public class JettyContainerService extends AbstractContainerService implements ContainerServiceEE8 {
 
   private static final Logger log = Logger.getLogger(JettyContainerService.class.getName());
